style(css): Reduce padding-right on content-body and header
The changes made in this commit are focused on reducing the padding-right on the `.content-body` and `.header` elements in the CSS file. This was done to improve the layout and responsiveness of the application, ensuring that the content is not pushed too far to the right, especially on smaller screens.

The specific changes are:

1. In the `.content-body` class, the `padding-right` value is changed from `5.625rem` to `0`.
2. In the `.header` class, the `padding-right` value is changed from `100px` to `0`.

These changes will help the content and header elements to be more closely aligned on the page, providing a more visually appealing and responsive layout.